requirements.txt
python = 3.6
Package              Version   
-------------------- ----------
absl-py              0.9.0     
astor                0.8.1     
certifi              2020.4.5.1
cffi                 1.14.0    
chardet              3.0.4     
cloudpickle          1.3.0     
cycler               0.10.0    
Cython               0.29.20   
decorator            4.4.2     
future               0.18.2    
gast                 0.2.2     
google-pasta         0.2.0     
grpcio               1.29.0    
h5py                 2.10.0    
horovod              0.19.1    
idna                 2.9       
imageio              2.8.0     
importlib-metadata   1.6.1     
joblib               0.15.1    
Keras-Applications   1.0.8     
Keras-Preprocessing  1.1.2     
kiwisolver           1.2.0     
llvmlite             0.33.0    
Markdown             3.2.2     
matplotlib           3.2.1     
natsort              5.4.1     
networkx             2.4       
numba                0.50.1    
numpy                1.18.2    
opencv-python        3.4.3.18  
opt-einsum           3.2.1     
pandas               1.0.4     
Pillow               7.1.2     
pip                  20.0.2    
protobuf             3.12.2    
psutil               5.7.0     
pycparser            2.20      
pydicom              2.0.0     
pyparsing            2.4.7     
python-dateutil      2.8.1     
pytz                 2020.1    
PyWavelets           1.1.1     
PyYAML               5.3.1     
requests             2.23.0    
scikit-image         0.14.5    
scikit-learn         0.23.1    
scipy                1.4.1     
setuptools           47.1.1    
SimpleITK            1.2.4     
six                  1.14.0    
tensorboard          1.15.0    
tensorboardX         2.1       
tensorflow-estimator 1.15.1    
tensorflow-gpu       1.15.0    
tensorlayer          1.5.0     
termcolor            1.1.0     
threadpoolctl        2.1.0     
tifffile             2020.6.3  
torch                1.4.0     
torchsummary         1.5.1     
torchvision          0.5.0     
tornado              6.0.4     
tqdm                 4.46.1    
urllib3              1.25.8    
Werkzeug             1.0.1     
wheel                0.34.2    
wrapt                1.12.1    
xlrd                 1.2.0     
zipp                 3.1.0     